% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mncitypop}
\alias{mncitypop}
\title{mncitypop}
\format{
\tabular{ll}{
\bold{Variable} \tab \bold{Description} \cr
County.Within         \tab County that the city or township exists within\cr
City.or.Township      \tab City or township name                         \cr
Year                  \tab Year of the estimate                          \cr
Total.Population      \tab Total population estimate                     \cr
Households            \tab Number of households                          \cr
Persons.Per.Household \tab Persons per household                         \cr
Home.County           \tab County of the city or town (in common format) \cr
City.Town             \tab City or town (in common format)
}
}
\source{
\url{https://mn.gov/admin/demography/data-by-topic/population-data/our-estimates/}
}
\usage{
mncitypop
}
\description{
Population estimates for each city/town in Minnesota, from the State Demographer
}
\examples{
data(mncitypop)
}
\keyword{datasets}
