% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mndemog-package.R
\docType{package}
\name{mndemog-package}
\alias{mndemog}
\alias{mndemog-package}
\title{mndemog: Data from the Minnesota State Demographers Office}
\description{
This package provides the mncitypop dataset a historical set of population data for all cities in Minnesota.
}
\author{
\strong{Maintainer}: Ryan Dunlap \email{163809015+ryadun@users.noreply.github.com}

}
\keyword{internal}
